added data generator for nces
