+++ nces_data_generator/generate_data.py
@@ -1,5 +1,5 @@
 import argparse, json
-from ontolearn.nces_utils import RDFTriples, KB2Data
+from helper_classes import RDFTriples, KB2Data
 parser = argparse.ArgumentParser()
 
 parser.add_argument('--kbs', type=str, nargs='+', default=['carcinogenesis'], help='Knowledge base name')
@@ -13,9 +13,9 @@ parser.add_argument('--rho', type=str, default='ExpressRefinement', choices=['Ex
 args = parser.parse_args()
 
 for kb in args.kbs:
-    triples = RDFTriples(source_kg_path=f'./NCESData/{kb}/{kb}.owl')
+    triples = RDFTriples(source_kg_path=f'../NCESData/{kb}/{kb}.owl')
     triples.export_triples()
-    with open(f'./NCESData/{kb}/data_generation_settings.json', "w") as setting:
+    with open(f'../NCESData/{kb}/data_generation_settings.json', "w") as setting:
         json.dump(vars(args), setting)
-    DataGen = KB2Data(path=f'./NCESData/{kb}/{kb}.owl', rho_name=args.rho, depth=args.depth, max_child_length=args.max_child_len, refinement_expressivity=args.refinement_expressivity, downsample_refinements=True, num_rand_samples=args.num_rand_samples, min_num_pos_examples=1, max_num_pos_examples=args.max_num_pos_examples)
+    DataGen = KB2Data(path=f'../NCESData/{kb}/{kb}.owl', rho_name=args.rho, depth=args.depth, max_child_length=args.max_child_len, refinement_expressivity=args.refinement_expressivity, downsample_refinements=True, num_rand_samples=args.num_rand_samples, min_num_pos_examples=1, max_num_pos_examples=args.max_num_pos_examples)
     DataGen.generate_descriptions().save_data()
